Notify author when topic replied

+++ database/migrations/2018_11_16_113242_create_replies_table.php
@@ -7,17 +7,21 @@ class CreateRepliesTable extends Migration
 {
 	public function up()
 	{
-		Schema::create('replies', function(Blueprint $table) {
-            $table->increments('id');
-            $table->integer('topic_id')->unsigned()->default(0)->index();
-            $table->integer('user_id')->unsigned()->default(0)->index();
-            $table->text('content');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('replies')) {
+            Schema::create('replies', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('topic_id')->unsigned()->default(0)->index();
+                $table->integer('user_id')->unsigned()->default(0)->index();
+                $table->text('content');
+                $table->timestamps();
+            });
+        }
 	}
 
 	public function down()
 	{
-		Schema::drop('replies');
+        if (Schema::hasTable('replies')) {
+            Schema::drop('replies');
+        }
 	}
 }
